+++ NEWS.md
@@ -1,4 +1,6 @@
-# Version 1.0.9000.9040
+# Version 1.0.9000.9041
+* Warning message not printed anymore by `nextBest` methods when `doselimit` not
+  specified.
 * Set prototype for `target = 0.3` for `NextBestMinDist` class.
 * Added new customized `checkmate` functions for probability values checking.
 * Created one `update` method for `ModelPseudo` and removed specific methods for